requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
Werkzeug==3.0.1
mysql-connector-python==8.3.0
python-dotenv==1.0.1
Pillow==10.2.0
APScheduler==3.10.4
WTForms==3.1.2
email-validator==2.1.0.post1